changing structure of the outliers generators and adding hypercube sampling
